fix: emulation image tag
