+++ samcli/commands/local/lib/local_lambda.py
@@ -2,32 +2,30 @@
 Implementation of Local Lambda runner
 """
 
-import os
 import logging
-from typing import Any, Dict, Optional, cast, List
-import boto3
+import os
+from typing import Any, Dict, Optional, cast
 
+import boto3
 from botocore.credentials import Credentials
 
 from samcli.commands.local.lib.debug_context import DebugContext
-from samcli.lib.providers.sam_function_provider import SamFunctionProvider
-from samcli.lib.utils.codeuri import resolve_code_path
-from samcli.lib.utils.packagetype import ZIP, IMAGE
-from samcli.lib.utils.stream_writer import StreamWriter
-from samcli.local.docker.container import ContainerResponseException
-from samcli.local.lambdafn.env_vars import EnvironmentVariables
-from samcli.local.lambdafn.config import FunctionConfig
-from samcli.local.lambdafn.exceptions import FunctionNotFound
 from samcli.commands.local.lib.exceptions import (
     OverridesNotWellDefinedError,
     NoPrivilegeException,
     InvalidIntermediateImageError,
-    UnsupportedRuntimeArchitectureError,
 )
 from samcli.lib.providers.provider import Function
+from samcli.lib.providers.sam_function_provider import SamFunctionProvider
+from samcli.lib.utils.architecture import validate_architecture_runtime
+from samcli.lib.utils.codeuri import resolve_code_path
+from samcli.lib.utils.packagetype import ZIP, IMAGE
+from samcli.lib.utils.stream_writer import StreamWriter
+from samcli.local.docker.container import ContainerResponseException
+from samcli.local.lambdafn.config import FunctionConfig
+from samcli.local.lambdafn.env_vars import EnvironmentVariables
+from samcli.local.lambdafn.exceptions import FunctionNotFound
 from samcli.local.lambdafn.runtime import LambdaRuntime
-from samcli.lib.utils.architecture import X86_64, ARM64
-
 
 LOG = logging.getLogger(__name__)
 
@@ -39,27 +37,6 @@ class LocalLambdaRunner:
     """
 
     MAX_DEBUG_TIMEOUT = 36000  # 10 hours in seconds
-
-    SUPPORTED_RUNTIMES: Dict[str, List[str]] = {
-        "nodejs10.x": [X86_64],
-        "nodejs12.x": [ARM64, X86_64],
-        "nodejs14.x": [ARM64, X86_64],
-        "python2.7": [X86_64],
-        "python3.6": [X86_64],
-        "python3.7": [X86_64],
-        "python3.8": [ARM64, X86_64],
-        "python3.9": [ARM64, X86_64],
-        "ruby2.5": [X86_64],
-        "ruby2.7": [ARM64, X86_64],
-        "java8": [X86_64],
-        "java8.al2": [ARM64, X86_64],
-        "java11": [ARM64, X86_64],
-        "go1.x": [X86_64],
-        "dotnetcore2.1": [X86_64],
-        "dotnetcore3.1": [ARM64, X86_64],
-        "provided": [X86_64],
-        "provided.al2": [ARM64, X86_64],
-    }
 
     def __init__(
         self,
@@ -151,7 +128,7 @@ class LocalLambdaRunner:
                 )
             LOG.info("Invoking Container created from %s", function.imageuri)
 
-        self.validate_architecture_runtime(function)
+        validate_architecture_runtime(function)
 
         config = self.get_invoke_config(function)
 
@@ -348,27 +325,3 @@ class LocalLambdaRunner:
             result["sessiontoken"] = creds.token
 
         return result
-
-    def validate_architecture_runtime(self, function: Function) -> None:
-        """
-        Validates that a function runtime and architecture are compatible for invoking
-
-        Parameters
-        ----------
-        function : samcli.commands.local.lib.provider.Function
-            Lambda function
-
-        Raises
-        ------
-        samcli.commands.local.lib.exceptions.UnsupportedRuntimeArchitectureError
-            If the runtime and architecture are not compatible
-        """
-        if function.packagetype == IMAGE:
-            return
-
-        runtime_architectures = self.SUPPORTED_RUNTIMES.get(cast(str, function.runtime), [])
-
-        if function.architectures and function.architectures[0] not in runtime_architectures:
-            raise UnsupportedRuntimeArchitectureError(
-                f"Runtime {function.runtime} is not supported on '{function.architectures[0]}' architecture"
-            )
